backend/ai-service/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0
langchain>=0.2.0,<1.0.0
langchain-google-genai>=1.0.0,<3.0.0
langchain-huggingface>=0.0.3,<1.0.0
langchain-community>=0.2.0,<1.0.0
faiss-cpu>=1.7.0,<2.0.0
sentence-transformers>=2.6.0,<4.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
numpy>=1.26.0,<3.0.0
scikit-learn>=1.4.0,<2.0.0
PyJWT>=2.8.0,<3.0.0
langchain-text-splitters>=0.2.0,<1.0.0
prophet>=1.1.5
pandas>=2.2.0
asyncpg>=0.29.0
pulp>=2.7.0
scipy>=1.11.0
